% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offtake-package.R
\docType{package}
\name{offtake-package}
\alias{offtake}
\alias{offtake-package}
\title{offtake: Sustainability Indices and Models for Wildlife Harvest and Hunting}
\description{
Toolkit for assessing the sustainability of wildlife exploitation (bushmeat hunting, recreational hunting and fisheries). It implements the main index-based and model-based approaches catalogued by Adounke et al. (2026) \doi{10.1016/j.biocon.2025.111606} and reviewed by Weinbaum et al. (2013) \doi{10.1111/ele.12008}. Indices compare a harvest or abundance metric between hunted and reference sites: age structure comparison ('asc'), hunting yield comparison ('hyco') and population density comparison ('pdc'). Models compare observed offtake with an estimated maximum sustainable harvest: the production model of Robinson and Redford (1991) ('pro'), potential biological removal of Wade (1998) \doi{10.1111/j.1748-7692.1998.tb00688.x} ('pbr'), the maximum sustainable yield of the logistic model ('msy'), the spatially explicit source-sink model of Levi et al. (2009, 2011) ('biode') and a stochastic extension of PBR in the spirit of Manlik et al. (2022) \doi{10.1111/cobi.13897} ('samse').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stangandaho/offtake}
  \item \url{https://stangandaho.github.io/offtake/}
  \item Report bugs at \url{https://github.com/stangandaho/offtake/issues}
}

}
\author{
\strong{Maintainer}: Stanislas Gandaho \email{stangandaho@gmail.com} (\href{https://orcid.org/0009-0003-4969-1252}{ORCID})

}
\keyword{internal}
